fix(pivot): Apply filteringExpressionTree in grid to reflect it in ESF
