+++ projects/igniteui-angular/src/lib/grids/pivot-grid/pivot-grid.pipes.ts
@@ -137,16 +137,8 @@ export class IgxPivotGridFilterPipe implements PipeTransform {
         advancedExpressionsTree: IFilteringExpressionsTree,
         _filterPipeTrigger: number,
         _pipeTrigger: number): any[] {
-        const allDimensions = config.rows.concat(config.columns).concat(config.filters).filter(x => x !== null && x !== undefined);
-        const enabledDimensions = allDimensions.filter(x => x && x.enabled);
+        const expressionsTree = PivotUtil.buildExpressionTree(config);
 
-        const expressionsTree = new FilteringExpressionsTree(FilteringLogic.And);
-        // add expression trees from all filters
-        PivotUtil.flatten(enabledDimensions).forEach((x: IPivotDimension) => {
-            if (x.filter && x.filter.filteringOperands) {
-                expressionsTree.filteringOperands.push(...x.filter.filteringOperands);
-            }
-        });
         const state = {
             expressionsTree,
             strategy: filterStrategy || new DimensionValuesFilteringStrategy(),
